backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.116.1",
    "openai>=1.98.0",
    "uvicorn>=0.34.0",
    "aiortc>=1.13.0",
]
